corpus/benign/mcp/gh-datawiz168-mcp-snowflake-service/pyproject.toml
# Python project configuration / Python 项目配置
[project]
name = "mcp-server-snowflake"
version = "0.1.0"
description = "MCP server for interacting with Snowflake databases / 用于与 Snowflake 数据库交互的 MCP 服务器"
readme = "README.md"
requires-python = ">=3.10"  # Required Python version / 所需的 Python 版本
dependencies = [  # Project dependencies / 项目依赖项
    "mcp>=1.0.0",                    # MCP SDK / MCP 软件开发工具包
    "snowflake-connector-python",     # Snowflake connector / Snowflake 连接器
    "python-dotenv"                   # Environment variable management / 环境变量管理
]

# Build system configuration / 构建系统配置
[build-system]
requires = ["hatchling"]  # Build tool requirement / 构建工具要求
build-backend = "hatchling.build"  # Build backend / 构建后端

# Build configuration / 构建配置
[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_snowflake"]  # Package location / 包位置

# Entry points / 入口点配置
[project.scripts]
mcp-server-snowflake = "mcp_server_snowflake.server:main"  # Command line entry / 命令行入口